Mudando o nome da classe Contato
