+++ src/main/java/br/com/fiap/NaMastreta/repository/ContatosRepository.java
@@ -2,7 +2,7 @@ package br.com.fiap.namastreta.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.fiap.namastreta.models.Contatos;
+import br.com.fiap.namastreta.models.Contato;
 
-public interface ContatosRepository extends JpaRepository<Contatos, Long> {
+public interface ContatosRepository extends JpaRepository<Contato, Long> {
 }
